pyproject.toml
[project]
name = "fantasy-coach"
version = "0.1.0"
description = "NRL match predictions: scraper, model, and API."
readme = "README.md"
requires-python = ">=3.12"
license = { text = "MIT" }
authors = [{ name = "Josh Lopez" }]
dependencies = [
    "fastapi>=0.115",
    "uvicorn[standard]>=0.32",
    "httpx>=0.27",
    "pydantic>=2.9",
    "numpy>=2.0",
    "scikit-learn>=1.5",
    "joblib>=1.4",
    "openpyxl>=3.1.5",
    "google-cloud-firestore>=2.19",
    "google-cloud-storage>=2.18",
    "firebase-admin>=6.5",
    "google-auth>=2.30",
    "xgboost>=3.2.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/fantasy_coach"]

[dependency-groups]
dev = [
    "pytest>=8.3",
    "pytest-cov>=5.0",
    "respx>=0.21",
    "ruff>=0.7",
    "pre-commit>=4.0",
]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "W", "I", "N", "UP", "B", "SIM"]

[tool.ruff.lint.per-file-ignores]
# sklearn convention: capital X for the feature matrix, capital C for regularisation.
"src/fantasy_coach/models/logistic.py" = ["N803", "N806"]
"src/fantasy_coach/models/calibration.py" = ["N803", "N806"]
"src/fantasy_coach/models/xgboost_model.py" = ["N803", "N806"]
"src/fantasy_coach/models/ensemble.py" = ["N803", "N806"]
"src/fantasy_coach/models/loader.py" = ["N803", "N806"]
"src/fantasy_coach/models/promotion.py" = ["N803", "N806"]
"src/fantasy_coach/drift.py" = ["N803", "N806"]
"src/fantasy_coach/retrain.py" = ["N803", "N806"]
"tests/test_ensemble.py" = ["N806"]
"tests/test_predictions.py" = ["N806"]
"src/fantasy_coach/evaluation/predictors.py" = ["N806"]
"src/fantasy_coach/feature_engineering.py" = ["N806"]
"tests/test_logistic.py" = ["N806"]
"tests/test_calibration.py" = ["N806"]
"tests/test_xgboost_model.py" = ["N806"]
# Fake Model-protocol implementations mirror sklearn naming (capital X).
"tests/test_promotion.py" = ["N803"]
"tests/test_retrain.py" = ["N803"]
# JSON API spec requires camelCase field names in the response schema.
"src/fantasy_coach/predictions.py" = ["N815"]
"src/fantasy_coach/app.py" = ["N815"]

[tool.pytest.ini_options]
addopts = "-ra -q --strict-markers"
testpaths = ["tests"]
